fix non-superuser test
