+++ codecept/tests/non-superuser/_log-in.js
@@ -1,0 +1,16 @@
+require('dotenv').config();
+const { generateToken } = require('authenticator');
+const test_user = {
+  username: process.env.TEST_USER_USERNAME,
+  email: process.env.TEST_USER_EMAIL,
+  key: process.env.TEST_USER_KEY,
+  password: process.env.TEST_USER_PASSWORD
+};
+const formatted_token = generateToken(test_user.key);
+
+
+Feature('Log in');
+
+Scenario('Log in', (I) => {
+  I.login(test_user, formatted_token);
+});
